Rename engine.condition -> engine.evaluate
