+++ src/engine/index.ts
@@ -1,4 +1,4 @@
-import { evaluateCondition } from './condition';
+import { evaluateCondition } from './evaluate';
 import { translate } from './translate';
 import { PageView, EngineCondition, AudienceDefinitionFilter } from '../../types';
